Bring LayoutRefreshController back under the file-length limit
LayoutRefreshController.swift sat at exactly 4000 lines, the file_length
error threshold, so the floating-window admission work pushed it five lines
over and turned `make verify` red on main. Generalising
deferTilingAdmissionIfNeeded into deferAdmissionIfNeeded also gave it a sixth
parameter, one past the function_parameter_count warning threshold.

Move the three monitor-geometry helpers - backingScale, preferredHideSide and
preferredHideSides - into a LayoutRefreshController+MonitorGeometry extension.
They are pure geometry over a monitor list with no dependency on the
controller's refresh state, so they carry their own weight in a separate file
and leave 46 lines of headroom. preferredHideSides widens from private to
internal because its remaining callers are now in the other file.

Collapse deferAdmissionIfNeeded's pid and windowId parameters into the
WindowToken all three call sites already had, which drops it back to five.
