+++ Sources/OmniWM/Core/Controller/LayoutRefreshController+MonitorGeometry.swift
@@ -1,0 +1,60 @@
+// SPDX-License-Identifier: GPL-2.0-only
+// Copyright (C) 2026 BarutSRB — https://github.com/BarutSRB/OmniWM
+
+import AppKit
+import Foundation
+
+@MainActor
+extension LayoutRefreshController {
+    /// Internal rather than private: `LayoutRefreshController` itself still calls this, and the
+    /// two types now live in different files.
+    func preferredHideSides(for monitors: [Monitor]) -> [Monitor.ID: HideSide] {
+        let important = 10
+        var preferredSides: [Monitor.ID: HideSide] = [:]
+
+        for monitor in monitors {
+            let monitorFrame = monitor.frame
+            let xOff = monitorFrame.width * 0.1
+            let yOff = monitorFrame.height * 0.1
+
+            let bottomRight = CGPoint(x: monitorFrame.maxX, y: monitorFrame.minY)
+            let bottomLeft = CGPoint(x: monitorFrame.minX, y: monitorFrame.minY)
+
+            let rightPoints = [
+                CGPoint(x: bottomRight.x + 2, y: bottomRight.y - yOff),
+                CGPoint(x: bottomRight.x - xOff, y: bottomRight.y + 2),
+                CGPoint(x: bottomRight.x + 2, y: bottomRight.y + 2)
+            ]
+
+            let leftPoints = [
+                CGPoint(x: bottomLeft.x - 2, y: bottomLeft.y - yOff),
+                CGPoint(x: bottomLeft.x + xOff, y: bottomLeft.y + 2),
+                CGPoint(x: bottomLeft.x - 2, y: bottomLeft.y + 2)
+            ]
+
+            func sideScore(_ points: [CGPoint]) -> Int {
+                monitors.reduce(0) { partial, other in
+                    let c1 = other.frame.contains(points[0]) ? 1 : 0
+                    let c2 = other.frame.contains(points[1]) ? 1 : 0
+                    let c3 = other.frame.contains(points[2]) ? 1 : 0
+                    return partial + c1 + c2 + important * c3
+                }
+            }
+
+            let leftScore = sideScore(leftPoints)
+            let rightScore = sideScore(rightPoints)
+            preferredSides[monitor.id] = leftScore < rightScore ? .left : .right
+        }
+
+        return preferredSides
+    }
+
+    func preferredHideSide(for monitor: Monitor) -> HideSide {
+        guard let controller else { return .right }
+        return preferredHideSides(for: controller.workspaceManager.monitors)[monitor.id] ?? .right
+    }
+
+    func backingScale(for monitor: Monitor) -> CGFloat {
+        NSScreen.screens.first(where: { $0.displayId == monitor.displayId })?.backingScaleFactor ?? 2.0
+    }
+}
